onist/project/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
httpx==0.27.0
pydantic==2.8.2
python-dotenv==1.0.1
beautifulsoup4==4.12.3
lxml==5.2.2
rapidfuzz==3.9.6
phonenumberslite==8.13.42
email-validator==2.2.0
Pillow==10.4.0
ImageHash==4.3.1